LayoutTests/fast/events/pointerleave-after-ancestor-removal-expected.txt

Verifies pointerleave fires on the retargeted ancestor after subtree removal.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS containerLeaveReceived is true
PASS successfullyParsed is true

TEST COMPLETE
Other
